gradio/requirements.txt
gradio
spaces
numpy
transformers
torch
rdkit
pillow
